kernel: do not delete a hard expired sa from kernel
A hard expired SA is already gone. Do not send delete request to the kernel for this
xfrm SA.

When the traffic is symetric, both SAs will expire one after the other.
While procssing the first hard expired sA the second one has alreday
expired in the kernel, and netlink message is queued for the pluto.
It is likely a corner case in real life, because traffic will not be
symetric.